update design in home coponeent
