--- client/src/components/Card/Card.styles.css
@@ -1,99 +1,0 @@
-.card {
-  text-decoration: none;
-  color: black;
-  position: relative;
-  width: 320px;
-  height: 500px;
-  background: #780000;
-  border-radius: 20px;
-  overflow: hidden;
-}
-.card:visited{
-  color: inherit
-}
-
-.card::before {
-  content: "";
-  position: absolute;
-  top: -50%;
-  width: 100%;
-  height: 100%;
-  background: #003049;
-  transform: skewY(345deg);
-  transition: 0.5s;
-}
-
-.card:hover::before {
-  top: -70%;
-  transform: skewY(390deg);
-}
-.card::after {
-  content: "e-Geek Store";
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  margin-left: 5px;
-  font-weight: 600;
-  font-size: 3.5em;
-  color: rgba(0, 0, 0, 0.1);
-}
-.imgBox {
-  position: relative;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-top: 20px;
-  z-index: 1;
-}
-
-.contentBox {
-  position: relative;
-  padding: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  z-index: 2;
-  color: white
-}
-
-.contentBox h2 {
-  font-size: 16px;
-  text-align: center;
-  color: white;
-  font-weight: 500;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-}
-
-.contentBox p {
-  font-size: 16px;
-  color: white;
-  font-weight: 700;
-  letter-spacing: 1px;
-}
-
-.buy {
-  position: relative;
-  top: 100px;
-  opacity: 1;
-  padding: 10px 30px;
-  color: white;
-  text-decoration: none;
-  background: #c1121f;
-  border-radius: 30px;
-  text-transform: uppercase;
-  letter-spacing: 1px;
-  transition: 0.5s;
-}
-
-.card:hover .buy {
-  top: 0;
-  opacity: 1;
-}
-
-.mouse {
-  height: 300px;
-  width: auto;
-}
