--- client/src/components/Cards/Cards.styles.css
@@ -1,7 +1,0 @@
-.container {
-  width: 85vw;
-  margin: 0 auto;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
-  grid-gap: 20px;
-}
