+++ client/src/views/Admin/Admin.styles.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Permanent+Marker&display=swap');
+/*@import url('https://fonts.googleapis.com/css2?family=Permanent+Marker&display=swap');
 
 .admin-container {
     max-width: 600px;
@@ -30,4 +30,4 @@
   .nav-link:hover {
     background-color: #ddd;
   }
-  
+  */
