+++ client/src/views/Home/Home.jsx
@@ -1,16 +1,15 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import InfiniteScroll from "react-infinite-scroll-component";
+
 import NavBar from "../../components/NavBar/NavBar";
 import Cards from "../../components/Cards/Cards";
 import Filters from "../../components/Filters/Filters";
-import FiltersTest from "../../components/Filters/FiltersTest";
 
-import Loading from "../../components/Loading/Loading";
 import { getProducts, resetHomeProducts } from "../../redux/actions";
-import InfiniteScroll from "react-infinite-scroll-component";
 import { fetchProducts } from "../../components/Pagination/Pagination";
 
-const Home = () => {
+export default function Home() {
   const dispatch = useDispatch();
   const allProducts = useSelector((state) => state.allProducts);
   const currentPage = useSelector((state) => state.currentPage);
@@ -19,13 +18,13 @@ const Home = () => {
 
   const loadMoreProducts = async () => {
     try {
-      dispatch({ type: "LOADING_TRUE" }); // Puedes manejar este tipo de acción según tu implementación
+      dispatch({ type: "LOADING_TRUE" });
       const data = await fetchProducts(currentPage + 1);
-      dispatch({ type: "LOADING_FALSE" }); // Puedes manejar este tipo de acción según tu implementación
-      dispatch({ type: "GET_PRODUCTS_SUCCESS", payload: data }); // Puedes manejar este tipo de acción según tu implementación
-    } catch (error) {
-      console.error("Error al cargar más productos:", error.message);
-      dispatch({ type: "LOADING_FALSE" }); // Puedes manejar este tipo de acción según tu implementación
+      dispatch({ type: "LOADING_FALSE" });
+      dispatch({ type: "GET_PRODUCTS_SUCCESS", payload: data });
+    } catch (err) {
+      console.error("Error:", err.message);
+      dispatch({ type: "LOADING_FALSE" });
     }
   };
 
@@ -39,22 +38,20 @@ const Home = () => {
   return (
     <div>
       <NavBar />
-      <FiltersTest />
-      <InfiniteScroll
-        dataLength={allProducts.length}
-        next={loadMoreProducts}
-        hasMore={!loading}
-        loader={<Loading />}
-      >
-        {filtered.length > 0 ? (
-          <Cards allProducts={filtered} />
-        ) : (
-          <Cards allProducts={allProducts} />
-        )}
-      </InfiniteScroll>
-      {loading && <p>Cargando...</p>}
+      <Filters>
+        <InfiniteScroll
+          dataLength={allProducts.length}
+          next={loadMoreProducts}
+          hasMore={!loading}
+          loader={""}
+        >
+          {filtered.length > 0 ? (
+            <Cards allProducts={filtered} />
+          ) : (
+            <Cards allProducts={allProducts} />
+          )}
+        </InfiniteScroll>
+      </Filters>
     </div>
   );
-};
-
-export default Home;
+}
