+++ client/src/views/ShippingForm/ShippingForm.css
@@ -1,4 +1,4 @@
-.containerForm{
+/*.containerForm{
     
     width: 600px;
     height: 800px;
@@ -84,3 +84,4 @@ p{
   color: red;
 }
 
+*/
