+++ client/src/views/UserForm/UserForm.styles.css
@@ -1,4 +1,4 @@
-form {
+/*form {
     margin: 20px;
     text-align: center; 
   }
@@ -41,4 +41,4 @@ form {
     background-color: rgb(200 30 30);
   }
   
-  
+  */
